Expose ChaincodeEventsRequest as interface in Java API
Consistency with other classes that implement behaviour.
